Welcome to the jungle
The purpose was only to add some test with jest to handle phone and link. But when
adding some library I encountered some errors because async key word was
deleted during transpilation. So I upggraded target to ES2022 and replaced babel by swc. always the
error... I replaced tsyringe by typedi, always the error... only
mongoose was using decorator, So I rewrited it to remove decorator. now
the error was about source map. I have added option to swc parseMap and
it is working now. during the process I have change maintainerData to
use controller and useState to have the same behavior than in
socialNetworkData

for the fonctionnal changes :

Can save phone and contact link for footer,
handle phone and contact as social network is maybe a bad idea. We will
see in the future