have a new block to handle invitation
it's a block with no possibility to add component. it works in two
steps.

the first is to set your name and surname. we search for you (no
diacritic and no case). if no result found, a modal is displayed to
explain that we don't find you

if you're found. A list of all members of you family is displayed with
the possibility to accept the invitation for the meal or the recetption.
a button to send the response allow to save the response.

the person has ot be set in database manually for now