can hide block until having a person invited to meal or reception
on block, it is now possible to say that a block is visible only when
invited to meal or to reception. So until we have a person logged in
with the right invitation, the block is hidden.

when log in, the person is set in cache and the page is reloaded. to do
so we set in cookie a session id, and we keep in cache the person