a lot of a thing are done here
eslint handles unused imports
we use storybook for some components
to do so, need to mock some controllers methods and avoid to load
dependencies container by default.
need to rework some components to give only used options (and not all
the component)

for the mock we use in the package.json the # alias, but it is in
conflict with alias of tsconfig.json. To have tsc command working, we
need to have a custom tsconfig.json to add the package.json alias
(otherwise it is not resolved)