have the possibility to add all block type
in a page we can have know a vertical or an horizontal block. And an
horizontal block can have a vertical block as child. to avoid unwanted
white space in padding, we remove it, if we need to we will add the
possibility in the admin to select the padding