Add a new page to display new orders
this screen displays all new orders and add button to validate or cancel
them.
A total line is displayed to show all product, their numbers and the
total price.

Need to adapt .swcrc to use it only for tests (otherwise production code
was broken)

Add an _Id attribute for customer to avoid a warning when sending the
object to the client side