have a component to login for admin pages
it is used for products, order and admin pages for now. to have something
pretty, we removed the layout with menu and footer which is useless
during setting up the app. the password is encoded before sending it to
the server, and then it is hashed. to set the password and the mail a
script is used on the server for now

the connection is not kept in cache. so every time the page is reloaded,
you need to login again. this will be improved in the next features